+++ packages/components/src/Form/Fields/FieldCheckbox/FieldCheckbox.story.tsx
@@ -39,10 +39,11 @@ Basic.args = {
   name: 'thumbsUp',
 }
 
-export const Detail = Template.bind({})
-Detail.args = {
+export const DetailDescription = Template.bind({})
+DetailDescription.args = {
   ...Basic.args,
-  detail: 'Some exciting details that describe the use of this.',
+  description: 'describe something here.',
+  detail: '4/20',
 }
 
 export const Checked = Template.bind({})
@@ -81,9 +82,9 @@ Error.args = {
   validationMessage: { message: 'This is an error', type: 'error' },
 }
 
-export const DetailError = Template.bind({})
-DetailError.args = {
-  ...Detail.args,
+export const DetailDescriptionError = Template.bind({})
+DetailDescriptionError.args = {
+  ...DetailDescription.args,
   validationMessage: { message: 'This is an error', type: 'error' },
 }