+++ packages/components/src/Form/Inputs/InputText/InputText.tsx
@@ -254,6 +254,7 @@ export const inputTextDisabled = css`
   background: ${(props) => props.theme.colors.ui1};
   color: ${(props) => props.theme.colors.text1};
   cursor: default;
+  -webkit-text-fill-color: ${(props) => props.theme.colors.text1};
   &:hover {
     border-color: ${(props) => props.theme.colors.ui2};
   }