+++ packages/components/src/Tree/stories/Tree.story.tsx
@@ -40,7 +40,7 @@ export default {
 }
 
 const Template: Story<TreeProps> = (args) => (
-  <Tree label="Orders" {...args}>
+  <Tree {...args} label="Orders">
     <Tree label="Orders" defaultOpen>
       <TreeItem>ID</TreeItem>
       <TreeItem>Status</TreeItem>